[XPath] Refactored evaluator code
